[REF] Netoyage de code sur le module maas_sale : - renommage fichier - suppression import non utilisé - suppression attribut redondante - reformat fichier python et xml - séparation fichier pour les fichiers contenants plusieurs class différents
