+++ maas-addons/maas_sale/data/ir_sequence.xml
@@ -1,15 +1,15 @@
 <?xml version="1.0" encoding="utf-8"?>
 <odoo>
-  <data noupdate="0">
+    <data noupdate="0">
 
-    <record id="seq_sale_operation" model="ir.sequence">
-      <field name="name">Operations</field>
-      <field name="code">sale.operation</field>
-      <field name="prefix">OP</field>
-      <field name="suffix"></field>
-      <field name="padding">3</field>
-      <field name="company_id" eval="False"/>
-    </record>
+        <record id="seq_sale_operation" model="ir.sequence">
+            <field name="name">Operations</field>
+            <field name="code">sale.operation</field>
+            <field name="prefix">OP</field>
+            <field name="suffix"></field>
+            <field name="padding">3</field>
+            <field name="company_id" eval="False" />
+        </record>
 
-  </data>
+    </data>
 </odoo>
