+++ maas-addons/maas_sale/models/__init__.py
@@ -1,13 +1,14 @@
 # -*- encoding: utf-8 -*-
 
 from . import sale_operation
-from . import product
+from . import product_template
 from . import res_partner
 from . import sale_campaign
 from . import package_upgrade
-from . import sale_subscription
+from . import sale_order
 from . import sale_campaign_action
-from . import product_pricelist
+from . import product_pricelist_item
 from . import res_config_settings
-from . import indication
+from . import indication_indication
+from . import partner_indication_infos
 from . import sale_operation_child
