+++ maas-addons/maas_sale/views/indication_views.xml
@@ -9,7 +9,7 @@
                     <field name="image" widget="image" class="oe_avatar" />
                     <div class="oe_title">
                         <h1>
-                            <field name="name" required="1"/>
+                            <field name="name" required="1" />
                         </h1>
                     </div>
                     <!--<group>
@@ -17,10 +17,10 @@
                     </group>-->
                     <field name="partner_info_ids" context="{'active_test':False}">
                         <tree>
-                            <field name="partner_id"/>
-                            <field name="quantity"/>
-                            <field name="sequence"/>
-                            <field name="active"/>
+                            <field name="partner_id" />
+                            <field name="quantity" />
+                            <field name="sequence" />
+                            <field name="active" />
                         </tree>
                     </field>
                 </sheet>
@@ -33,7 +33,7 @@
         <field name="model">indication.indication</field>
         <field name="arch" type="xml">
             <tree>
-                <field name="name"/>
+                <field name="name" />
             </tree>
         </field>
     </record>
@@ -45,19 +45,21 @@
             <form>
                 <sheet>
                     <div class="oe_button_box" name="button_box">
-                        <button name="toggle_active" type="object" class="oe_stat_button" icon="fa-archive">
-                            <field name="active" widget="boolean_button" options="{&quot;terminology&quot;: &quot;archive&quot;}"/>
+                        <button name="toggle_active" type="object" class="oe_stat_button"
+                            icon="fa-archive">
+                            <field name="active" widget="boolean_button"
+                                options="{&quot;terminology&quot;: &quot;archive&quot;}" />
                         </button>
                     </div>
                     <group>
                         <group>
-                            <field name="partner_id" required="1"/>
-                            <field name="quantity"/>
+                            <field name="partner_id" required="1" />
+                            <field name="quantity" />
                         </group>
                     </group>
                     <group>
                         <group>
-                            <field name="sequence"/>
+                            <field name="sequence" />
                         </group>
                     </group>
                 </sheet>
@@ -71,12 +73,6 @@
         <field name="res_model">indication.indication</field>
         <field name="view_mode">tree,form</field>
     </record>
-
-    <menuitem id="indication_menu"
-                  name="Indications"
-                  parent="sale_operation_menu_root"
-                  sequence="50"
-                  action="indication_action"/>
 
 
 </odoo>
