+++ maas-addons/maas_sale/views/res_partner_view.xml
@@ -5,15 +5,15 @@
         <record id="res_partner_view_form_inherit" model="ir.ui.view">
             <field name="name">res.partner.view.form.inherit</field>
             <field name="model">res.partner</field>
-            <field name="inherit_id" ref="base.view_partner_form"/>
+            <field name="inherit_id" ref="base.view_partner_form" />
             <field name="arch" type="xml">
                 <field name="vat" position="after">
-                    <field name="title_ref"/>
+                    <field name="title_ref" />
                 </field>
                 <group name="misc" position="after">
                     <group name="indication_grp" string="Indication">
-                        <field name="date_update_title"/>
-                        <field name="date_update_vars"/>
+                        <field name="date_update_title" />
+                        <field name="date_update_vars" />
                     </group>
                 </group>
             </field>
