+++ maas-addons/maas_sale/views/sale_campaign.xml
@@ -1,81 +1,85 @@
 <?xml version="1.0" encoding="utf-8"?>
 <odoo>
-  <data>
+    <data>
 
-    <record id="sale_campaign_view_form" model="ir.ui.view">
-      <field name="name">sale.campaign.view.form</field>
-      <field name="model">sale.campaign</field>
-      <field name="arch" type="xml">
-        <form string="Campaign">
-          <sheet>
-            <div class="oe_title">
-              <h2>
-                <field name="name" nolabel="1"/>
-              </h2>
-            </div>
-            <group>
-              <group>
-                <field name="partner_id" options="{'no_open': True, 'no_create': True}"/>
-              </group>
-            </group>
-            <group name="action_grp" string="Actions">
-              <field name="action_ids" nolabel="1" colspan="2">
-                <tree>
-                  <field name="name"/>
-                  <field name="campaign_id"/>
+        <record id="sale_campaign_view_form" model="ir.ui.view">
+            <field name="name">sale.campaign.view.form</field>
+            <field name="model">sale.campaign</field>
+            <field name="arch" type="xml">
+                <form string="Campaign">
+                    <sheet>
+                        <div class="oe_title">
+                            <h2>
+                                <field name="name" nolabel="1" />
+                            </h2>
+                        </div>
+                        <group>
+                            <group>
+                                <field name="partner_id"
+                                    options="{'no_open': True, 'no_create': True}" />
+                            </group>
+                        </group>
+                        <group name="action_grp" string="Actions">
+                            <field name="action_ids" nolabel="1" colspan="2">
+                                <tree>
+                                    <field name="name" />
+                                    <field name="campaign_id" />
+                                </tree>
+                            </field>
+                        </group>
+                        <group name="operation_grp" string="Operations">
+                            <field name="operation_ids" nolabel="1" colspan="2">
+                                <tree>
+                                    <field name="number" />
+                                    <field name="name" />
+                                    <field name="partner_id" />
+                                    <field name="date" />
+                                    <field name="qty_extracted" />
+                                    <field name="type" />
+                                    <field name="state" />
+                                </tree>
+                            </field>
+                        </group>
+                    </sheet>
+                </form>
+            </field>
+        </record>
+
+        <record id="sale_campaign_view_tree" model="ir.ui.view">
+            <field name="name">sale.campaign.view.tree</field>
+            <field name="model">sale.campaign</field>
+            <field name="arch" type="xml">
+                <tree string="Campaigns">
+                    <field name="name" />
+                    <field name="partner_id" />
+                    <field name="action_ids" />
+                    <field name="operation_ids" />
                 </tree>
-              </field>
-            </group>
-            <group name="operation_grp" string="Operations">
-              <field name="operation_ids" nolabel="1" colspan="2">
-                <tree>
-                  <field name="number"/>
-                  <field name="name"/>
-                  <field name="partner_id"/>
-                  <field name="date"/>
-                  <field name="qty_extracted"/>
-                  <field name="type"/>
-                  <field name="state"/>
-                </tree>
-              </field>
-            </group>
-          </sheet>
-        </form>
-      </field>
-    </record>
+            </field>
+        </record>
 
-    <record id="sale_campaign_view_tree" model="ir.ui.view">
-      <field name="name">sale.campaign.view.tree</field>
-      <field name="model">sale.campaign</field>
-      <field name="arch" type="xml">
-        <tree string="Campaigns">
-          <field name="name"/>
-          <field name="partner_id"/>
-          <field name="action_ids"/>
-          <field name="operation_ids"/>
-        </tree>
-      </field>
-    </record>
+        <record id="sale_campaign_search" model="ir.ui.view">
+            <field name="name">sale.campaign.search</field>
+            <field name="model">sale.campaign</field>
+            <field name="arch" type="xml">
+                <search string="Sale Campaign">
+                    <field name="operation_ids" filter_domain="[('name','ilike',self)]"
+                        string="Operation" />
+                    <field name="operation_ids" filter_domain="[('number','ilike',self)]"
+                        string="Num Operation" />
+                    <field name="operation_ids" filter_domain="[('partner_id.name','ilike',self)]"
+                        string="Client" />
+                </search>
+            </field>
+        </record>
 
-    <record id="sale_campaign_search" model="ir.ui.view">
-      <field name="name">sale.campaign.search</field>
-      <field name="model">sale.campaign</field>
-      <field name="arch" type="xml">
-        <search string="Sale Campaign">
-          <field name="operation_ids" filter_domain="[('name','ilike',self)]" string="Operation"/>
-          <field name="operation_ids" filter_domain="[('number','ilike',self)]" string="Num Operation"/>
-          <field name="operation_ids" filter_domain="[('partner_id.name','ilike',self)]" string="Client"/>
-        </search>
-      </field>
-    </record>
+        <record id="sale_campaign_action" model="ir.actions.act_window">
+            <field name="name">Campaigns</field>
+            <field name="type">ir.actions.act_window</field>
+            <field name="res_model">sale.campaign</field>
+            <field name="view_mode">tree,form</field>
+            <field name="search_view_id" ref="sale_campaign_search" />
+        </record>
 
-    <record id="sale_campaign_action" model="ir.actions.act_window">
-      <field name="name">Campaigns</field>
-      <field name="type">ir.actions.act_window</field>
-      <field name="res_model">sale.campaign</field>
-      <field name="view_mode">tree,form</field>
-      <field name="search_view_id" ref="sale_campaign_search"/>
-    </record>
-
-  </data>
+    </data>
 </odoo>
