+++ maas-addons/maas_sale/views/sale_campaign_action.xml
@@ -1,0 +1,64 @@
+<?xml version="1.0" encoding="utf-8"?>
+<odoo>
+    <data>
+
+        <record id="sale_campaign_action_view_form" model="ir.ui.view">
+            <field name="name">sale.campaign.action.view.form</field>
+            <field name="model">sale.campaign.action</field>
+            <field name="arch" type="xml">
+                <form string="Action">
+                    <sheet>
+                        <div class="oe_title">
+                            <h2>
+                                <field name="name" nolabel="1" />
+                            </h2>
+                        </div>
+                        <group>
+                            <field name="campaign_id" options="{'no_create': True}" />
+                        </group>
+                        <group name="operation_grp" string="Operations">
+                            <field name="operation_ids" nolabel="1" colspan="2" />
+                        </group>
+                    </sheet>
+                </form>
+            </field>
+        </record>
+
+        <record id="sale_campaign_action_view_tree" model="ir.ui.view">
+            <field name="name">sale.action.view.form</field>
+            <field name="model">sale.campaign.action</field>
+            <field name="arch" type="xml">
+                <tree string="Actions">
+                    <field name="name" />
+                    <field name="campaign_id" />
+                    <field name="operation_ids" />
+                </tree>
+            </field>
+        </record>
+
+        <record id="sale_campaign_action_search" model="ir.ui.view">
+            <field name="name">sale.campaign.action.search</field>
+            <field name="model">sale.campaign.action</field>
+            <field name="arch" type="xml">
+                <search string="Sale Campaign action">
+                    <field name="operation_ids" filter_domain="[('name','ilike',self)]"
+                        string="Operation" />
+                    <field name="operation_ids" filter_domain="[('number','ilike',self)]"
+                        string="Num Operation" />
+                    <field name="operation_ids" filter_domain="[('partner_id.name','ilike',self)]"
+                        string="Client" />
+                </search>
+            </field>
+        </record>
+
+        <record id="sale_campaign_action_action" model="ir.actions.act_window">
+            <field name="name">Actions</field>
+            <field name="type">ir.actions.act_window</field>
+            <field name="res_model">sale.campaign.action</field>
+            <field name="view_mode">tree,form</field>
+            <field name="view_mode">tree,form</field>
+            <field name="search_view_id" ref="sale_campaign_action_search" />
+        </record>
+
+    </data>
+</odoo>
