--- maas-addons/maas_sale/views/sale_order_views.xml
@@ -1,6 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<odoo>
-  <data>
-
-  </data>
-</odoo>
